try to deploy

+++ packages/contracts/src/systems/DigResourceSystem.sol
@@ -102,7 +102,7 @@ contract DigResourceSystem is System {
         );
         (uint256 remain, uint256 cache, uint256 difficulty) = getRemainAndCache(resourceId, digInfo.perlin);
         require(remain == digInfo.remain && cache == digInfo.cache, "remain value invalid");
-        uint256 powResult = keccak256(abi.encodePacked(digInfo.coordHash, digInfo.remain, digInfo.powNonce)); 
+        uint256 powResult = uint256(keccak256(abi.encodePacked(digInfo.coordHash, digInfo.remain, digInfo.powNonce))); 
         require(powResult / 16 	** (64 - difficulty) == 0, "pow value invalid");
         resourcePosition.set(resourceId, digInfo.coordHash);
         resourceMining.set(resourceId, ResourceMining({remain: remain-1, cache: cache+1}));
